Remove reconcile-cluster-roles and reconcile-cluster-role-bindings
This change removes the oc policy reconciliation commands.  These
commands were deprecated in v3.9 in favor of "oc auth reconcile."
Since the API master has done automatic reconciliation on server
start since v3.7, these commands are no longer necessary.
